refactor: move all note types to note-types.ts
