--- src/ts/types/pause.ts
@@ -1,1 +1,0 @@
-export type Pause = 'pause';
